requirements.txt
python-binance
TA-Lib
numpy
websocket_client
slackclient
django-environ
